backend/requirements.txt
langchain==0.2.11
langchain-openai==0.1.17
langchain-community==0.2.10
langchain-core==0.2.23
pydantic==2.8.2
pydantic-core==2.20.1
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
fastapi==0.111.1
uvicorn[standard]==0.30.3
httpx==0.25.0
beautifulsoup4==4.12.2
asyncpg==0.29.0
aiohttp==3.9.1
cachetools==5.3.2